requirements.txt
# Django==4.2.0
# Pillow==10.0.0
# opencv-python==4.8.0
# face-recognition==1.3.0
# dlib==19.24.0
# numpy==1.24.0
# requests==2.31.0
# psycopg2-binary==2.9.6
# python-dotenv==1.0.0
# gunicorn==21.2.0
# Django==4.2.0
# Pillow>=9.5,<11.0
# opencv-python==4.8.0.76
# face-recognition==1.3.0
# numpy>=1.23,<1.26
# requests==2.31.0
# psycopg2-binary==2.9.6
# python-dotenv==1.0.0
# gunicorn==21.2.0

# Core Django and Database
Django>=4.2.0
psycopg2-binary>=2.9.0  # or use SQLite

# Face Recognition (existing)
face-recognition>=1.3.0
pillow>=10.0.0

# Pill Recognition - Computer Vision
opencv-python>=4.8.0
numpy>=1.24.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep Learning (optional but recommended)
tensorflow>=2.13.0  # or tensorflow-cpu for CPU-only
keras>=2.13.0

# Image Processing
opencv-contrib-python>=4.8.0  # Additional OpenCV modules

# QR Code Generation
qrcode[pil]>=7.4.2

# Other
requests>=2.31.0
python-dotenv>=1.0.0